Add description fucntion
